Added channel & nickname keys.
Since IRC does not have case-sensitivity, and since the maps are case-sensitive, we had to normalize and generate a nickname & channel key.  The keys are simply normalized, lowercase versions of the nickname and channel name.
Added a timeout option for the setStatus method so that a message will disappear over time if included.
Added disconnect & reconnect messages.
Added closedByServer information to the disconnect event.